requirements.txt
chainlit
fastapi
uvicorn
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
chromadb>=0.5.0
pydantic>=2.6
pydantic-settings>=2.2
tiktoken
pypdf
docx2txt
sentence-transformers
urllib3>=2.5.0
google-cloud-aiplatform>=1.43.0
google-cloud-discoveryengine>=1.6.0
google-auth>=2.20.0
google-cloud-storage>=2.10.0
langchain-google-vertexai>=1.0.0
notion-client>=2.2.1
llama-index-core>=0.10.0
llama-index-llms-vertex>=0.1.0
llama-index-embeddings-vertex>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
pytest>=8.0
